Implement LDAP domains using last keystone host
To ensure that the domains can be created, we only execute
the tasks on the last member of the keystone_all group so
that the load balancer has at least one other host to work
against.

We also replace the OSA keystone module in the task with
the standard Ansible module, and delegate the task to the
service setup host.

Change-Id: I66ed21cdcf42d0c2012062c8cf74305fecbec312