Only do sshd enablement task on the first play host
We only need to do this task once in the play because it
delegates to all the hosts already. Currently it will do
the sshd enablement 9 times for 3 hosts.

We also add a comment to explain why this is being done
for future reference.

Change-Id: I201101f6d2bf733b375577b6af138272b9d0ce6b
(cherry picked from commit aefb58080337633e6d4c10bb2709c8d73ab9564d)