Add memcache flushing handler on db migrations
Flushing cache when schema changes occur to avoid drift between
what might be cached and what keystone queries expect.

Given that this is a workaround which is only usually required
in major upgrades (and should really be handled in process by
keystone itself), we make the execution optional so that it can
be set to happen only on major upgrades.

To prevent test breakage as happened in master, we combine the
two master patches into one for the backport:

https://review.openstack.org/608066
https://review.openstack.org/613256

Depends-On: https://review.openstack.org/613290
Change-Id: Ibf8f3dd60d6f3c446a14dc8228fa005f12fcc840
Closes-Bug: 1793389