Implement LDAP domains using last keystone host
To ensure that the domains can be created, we only execute
the tasks on the last member of the keystone_all group so
that the load balancer has at least one other host to work
against.

We also replace the OSA keystone module in the task with
the standard Ansible module, and delegate the task to the
service setup host.

Backport note:
This does not implement the changes to tasks/keystone_ldap_setup.yml
as was done for Rocky/Master because the appropriate changes to
support those have not been implemented. This is a minimal functional
backport to support fixing the bug.

Closes-Bug: #1791873
Change-Id: I66ed21cdcf42d0c2012062c8cf74305fecbec312
(cherry picked from commit 096ed19665dad45fe9b1c1790ba599d90ccf233c)