Make the memcache flush optional
The memcache flush implemented in https://review.openstack.org/608066
is actually a workaround. The implementation should really be
implemented more surgically in keystone itself. This has begun with
https://review.openstack.org/612686 but it is not complete. However,
something that's been made clear by the team is that this was only
required for the Newton->Ocata upgrade and has not been required since.

This workaround may be required again in the future, so instead of
removing it, we make the tool opt-in via a toggle which can then easily
be set when doing the appropriate major upgrade.

Related-Bug: 1793389
Change-Id: Ied0ce1e9877697bb627f784a0590f7c7e924479b