Revert "vars: Add missing netcat package"
This reverts commit 781835e7521b91ef813a1c95455ac9e91794a8ca.

The package is actually required on the memcached hosts, not
the keystone hosts. This helped make role tests pass because
in the role tests memcache and keystone are often on the same
containers.

The actual fix will be https://review.openstack.org/613099 which
ensures that netcat is installed on the memcache hosts instead.

Depends-On: https://review.openstack.org/613099
Change-Id: I55158c332a35e150f61541e8c2c1390e397d7d2c