Do not remove SSO callback template file
In distro installs, the SSO callback template file is created in
/etc/keystone. The 'Retrieve default configuration files from venv' task
copies it from there to a cache folder. If 'keystone_sp' is not provided
the file is being removed, which causes a subsequent run of the role to
fail.

Change-Id: I05df4c2a5f57c7437c008dd73a31b7eda2c3e281
(cherry picked from commit f66eca27d32fefc334d5b0f90c64f551e9b19116)