Ensure that LDAP config is deployed on all keystone hosts
In I66ed21cdcf42d0c2012062c8cf74305fecbec312 the condition meant
for the setup of the domain was mistakenly applied to all tasks
including laying down the template.

This patch moves the conditional which ensures the domain is setup
on the first host to the task in question to ensure that everything
is good and well with the world again.

Change-Id: Icb7c2556306d459534e6791f16c7013d0e9fcaf5
Closes-Bug: #1804827