Enable overriding the service setup host python interpreter
In order to enable the service setup host python interpreter to
be changed easily, we make it a variable. This will be useful
when someone sets the service setup host to be the utility
container, because we'll be able to set this var by default.

Change-Id: If33382404565dcb288e2b6794512f09c85e751d5
(cherry picked from commit 9573f2e714797674f33025d4c99b29cb4c7e5c82)