Correct notification driver
The notification driver setup was resulting in the driver and connection string
on the same line. This is caused by the case statement and how jinja formats
the template when a case statement is present. This change modifies how the
driver string is created using a ternary, which will eliminate the case
statement and render the value of the diver correctly.

Change-Id: I78293b35a30ea12da25fc11d7a5c33cdf2982e2d
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>