Update role for new source build process
The developer mode and venv download modes at the moment
no longer carry any meaning. This review changes this role to
do the equivalent of what developer_mode was all the time,
meaning that it always builds the venv and never requires
the repo server, but it will use a repo server when available.

As part of this, we move the source build out of its own file
because it's now a single task to include the venv build role.
This is just to make it easier to follow the code.

Change-Id: I53d1fb9ace5eb3e66ca8aef08603abf580acc066