handlers: check for default IP instead of 127.0.0.1
There is no host set for the handler that checks once the service
has gone up.  In certain scenario where you may be running all
metal with HAproxy on the same node, you would want Keystone to
bind to the primary IP only to avoid conflicting with HAproxy.

By doing this, you're no longer listening to 0.0.0.0 which includes
127.0.0.1 and this handler fails.  By switching the handler to hit
the primary IP, you're able to successfully run this task both in
the existing scenario, as well as cover scenario where you are
listening on the primary IP.

Change-Id: I1f519d74ec25abfe37d2952aac92a346ccbf48f6