Remove keystone directory on source installs
Current behavior only attempts to remove the keystone
directory from the first container and skips
additional containers past the first one.  This
caused upgrades to break as the configs were
still present in any additional containers.

This ensures the keystone directory is removed on
all keystone containers when the install method is
is source.

Change-Id: If588f9ed4bc5d0deeb2b9c1bbeea5e9eb5ce7c79