Fix custom policy configuration
Custom policies for Keystone are implemented completely different
from other roles, in a way that is broken from prior OSA branches.

On older and newer branches, keystone_policy_overrides is used to
provide a yaml dict of policy overrides using ansible vars. The
current implementation of keystone's policy distribution requires
that a base policy be configured in a flat file on the deployment
host at /etc/openstack_deploy/keystone/policy.json, to which policy
overrides are then applied and copied to the keystone hosts.

If the flat file does not exist, no policy overrides are applied
to Keystone.

This change modifies the policy distribution to work similarly to
other OSA roles in the rocky branch. No flat file is sourced from
/etc/openstack_deploy any longer, only the keystone_policy_overrides
var is used, and a policy.json is only deployed to Keystone
if this var is defined by the deployer.

Change-Id: I7f6b80acb27fb959ce6d68bd35d3cc5dccc86f9f