Remove references to writable LDAP from documentation
Keystone removed support for writable LDAP backends in Ocata. Prior to
that it was deprecated for several releases. This commit removes
references to those configuration options since they are silently
ignored. This cleans up the configuration files and doesn't give the
impression that functionality is still supported.

Relevant release notes that advertize the removal:

  https://docs.openstack.org/releasenotes/keystone/ocata.html#relnotes-11-0-0-origin-stable-ocata-other-notes

Change-Id: Id05247d004ee7d189dff3ec867a6ec11dfc40e9b