Delegate identity provider setup to service setup host
This patch delegates tasks setting up a keystone identity provider to the
keystone_service_setup_host becasue the ansible modules used require the
python-keystoneclient module. These tasks will fail when the keystone
(container) host is not the deploy host as python-keystoneclient will
be missing.

These tasks have recently been refactored in a previous patch [1] to use
the os_* ansible modules instead of the OSA specific keystone module.
This patch fixes a reported bug with the old code which is also present
after that refactoring.

[1] https://review.opendev.org/#/c/656397/

Change-Id: I16a61501a0578e87464e8bd1031af4270eac2d98
Closes-Bug: #1856165