Identity Providers support improvments
Identity providers can be created with specified domain
IDPs are linked to a keystone domain. Currently when we provision a new IDP
it gets created within a new autogenerated domain. Provisioners can now
give a domain_id in an IDP entry within the trusted_idp_list and the IDP
will be created within that domain.

Add IDP display_name to defaults
Allows operator a name different to the one used in the backend
to display to the user in the horizon Identity Provider dropdown.

Change-Id: Iaf9f1b9198f14c903f9801e0ce7da86b74d9c5bd
(cherry picked from commit 4c9642765e632d409c6d9028aae26f4514af7033)